+++ lib/tidefall/hash_map.ex
@@ -588,8 +588,9 @@ defmodule Tidefall.HashMap do
   ## Options
 
   Updatable options: `:processing_interval`, `:processing_timeout`,
-  `:processing_batch_size`. See [start options](`m:Tidefall.HashMap#module-start-options`)
-  for each option's semantics.
+  `:processing_batch_size`, `:drain_threshold`, `:drain_check_interval`. See
+  [start options](`m:Tidefall.HashMap#module-start-options`) for each option's
+  semantics.
 
   ## Examples